pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "helpertips"
version = "0.1.0"
description = "Captura sinais de futebol virtual do Telegram e armazena no PostgreSQL"
requires-python = ">=3.12"
dependencies = [
    "telethon~=1.42",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "dash>=4.1,<5",
    "dash-bootstrap-components>=2.0",
    "dash-ag-grid>=31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
target-version = "py312"
line-length = 150
exclude = [".claude", ".git", "__pycache__", "*.egg-info", "venv"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []
